Init of a hacked away Binary game for cljs.pro learning project.
This began as 4clojure ex's, then the beginnings of a whack-a-mole game
and then decided to switch again into a web Binary Number's Game.
The game is in a single file, asking to be logically namespaced.
There's a lot of progression q's I ask myself, with the intention of
just beginning the actual game in another repo.  But why now when
I can use this and just clean it up.